Add oracle parsing for clock advancement
Before, oracle entries were expected to refer to a built-in oracle with
link to datasworn, but e.g. rolling likelihood oracle to advance clocks
has a different format. Now it accounts for that, too.

Added util function to convert links to text, plus testing for that.
This might need some changes later to include few options to e.g.
extract the link itself, or make it an actual link.