Add status parameter to clock node parser
Clock node parser only handled the progress, but not status when the
clock was added or resolved / removed.

Handling both the resolved and removed option. Iron-Vault doc mentions
"removed" as status option, though that was changed to "resolved" in
https://github.com/iron-vault-plugin/iron-vault/commit/353cb75 back in
August 2024. So probably journals from before that time will have the
removed option used.