Rename node parser back to NodeParser
TemplateRegexNodeParsser feels a bit overkill of a name now that there
aren't any other types of parsers anymore (apart form the fallback one).
Might as well call it back to the original NodeParser. Yeah, could have
done this all in a single step, but hopes are that this way it will be
slightly let confusing if I ever need to revisit this in the future.