Use more flexible template file for clock nodes
There's two different scenarios in the clock node: status change of the
clock, and progress. Currently some hardcoded text is created based on
either case, leaving not much flexibility. There's Jinja in place, so
both cases could be easily handled on their own with that.

Moving the template to its own file and passing all available data to
it, with the template handling each case then based on that data.