Introduce Context class for parser data
Currently only the parent ElementTree object was of relevance for
parsing nodes, but in order to e.g. group rolls, rerolls and burning
momentum into a single move outcome, more data will be necessary to keep
around. Packing into a general `Context` object is the attempt for that.

Moving the parent ElementTree in there as first step, and also extending
it to a stack so different element levels and children can use the same
context object and the shared data within it.