Add roll outcome div to move nodes
Adds a div with the total roll result, including reroll and momentum
burn adjustments at the bottom of a move node - if rolls happened in it.

This is a good start but needs some improvements. Would be nicer to have
the result on top, and be able to style the whole move node based on its
final result.