Add support to disable node templates
Possibly not all the nodes are of interest in the HTML output, for
example the newly introduced `roll_result` node that adds an extra div
for the roll outcome to a move.

Setting the user template value now to an empty string will ignore
rendering it to HTML during parsing. All other activity happening during
parsing, i.e. regex matching and roll context handling, will still
happen regardless of the template state.