Use templates for the main mechanics block <div>
Getting rid of the last hard-coded, non-template block, the main block
that contains an iron-vault-mechanics block.

Instead of creating the `<div>` first and then pass it to `Context`
during initialization, `Context` itself is creating it now and attaches
it to the root element passed to its constructor.

Adds a template file and user template override option for it, but no
template arguments are handled. It's primarily to give a option for
custom styling.