Fully use templates for block parsers
BlockParsers collect nodes in a tempoarary `<div>` now, and during the
`finalize()` call replace it with their parsed template. This allows to
define templates with data that isn't available from the starting line
of the block, which is especially relevant for Move blocks.

Before, the Move block parser just tweaked the CSS classes based on the
roll outcome, but now all the roll outcome is available as variables in
its template. This allows a whole lot more flexibility with styling.