Rename template-related names and config keys
Changing extension config keys:
 - `templates` to `template_overrides`
 - `theme` to `themplate_path`

Internally, anything `user_templates` is renamed to simply `overrides`
as that's what those are now, and while it helps to set up themes, it's
really only a path to templates as alternative to the package-provided
templates, so that becomes simply `path` now.

To keep the naming consistent, the`UserTemplates` class was renamed to
`TemplateOverrides`.

Sorry if anyone is actually using this extension and has to tweak their
code now because of this :/