The return of the roll-result
Now that templates a lot easier and flexible to customize, I don't need
to decide whether I like the roll-result option or not. So why not have
it back in the node list.

This adds an extra roll-result node at the end of a move block, provided
there was a roll happening, and a roll-result template is available.